Allow chart levels to use one decimal place

+++ src/gameplay/chart_difficulty.h
@@ -5,6 +5,6 @@
 namespace chart_difficulty {
 
 float calculate_rating(const chart_data& data);
-int calculate_level(const chart_data& data);
+float calculate_level(const chart_data& data);
 
 }  // namespace chart_difficulty
